refactor(api): extract the shared usage view builder into usage_view.py
Move the 17 usage-view definitions (DashboardUsageRecord/View, _build_usage_view,
_usage_record_time, the record builders and their helpers) verbatim out of api.py
into a new data-layer module, so non-web consumers (usage_snapshot, plan_cost,
and through them the TUI) stop importing the web server module for pure data
logic. api.py re-imports every moved name and behaves exactly as before.

Zero behavior change: each moved definition is byte-identical to the original
(AST-verified), and api's attributes are the same objects usage_view defines.

Groundwork for retiring the HTML dashboard: the display layer can now shrink
without touching the shared usage/cost data path.
